# api/openapi.yaml
openapi: 3.0.3
info:
  title: Carpooling VE API
  description: |
    API para la plataforma de carpooling interurbano de Venezuela.
    Corredor inicial: Guatire ↔ Caracas.
    MVP — Mes 3.
  version: 0.1.0
  contact:
    name: Engineer
servers:
  - url: https://api.staging.carpooling.ve
    description: Staging
  - url: http://localhost:3000
    description: Local dev
security:
  - bearerAuth: []
...
paths:
...
  /users/{id}:
    get:
      tags: [Users]
      summary: Obtener usuario por ID
      parameters:
        - $ref: '#/components/parameters/UserId'
      responses:
        '200':
          description: Usuario encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
    patch:
      tags: [Users]
      summary: Actualizar perfil
      parameters:
        - $ref: '#/components/parameters/UserId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserUpdate'
      responses:
        '200':
          description: Usuario actualizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'

  # ─── DRIVERS ───
...
  /vehicles:
    post:
      tags: [Vehicles]
      summary: Registrar vehículo
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VehicleCreate'
      responses:
        '201':
          description: Vehículo registrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Vehicle'
    get:
      tags: [Vehicles]
      summary: Listar vehículos del conductor
      responses:
        '200':
          description: Lista de vehículos
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Vehicle'
  /vehicles/{id}:
    get:
      tags: [Vehicles]
      summary: Obtener vehículo por ID
      parameters:
        - $ref: '#/components/parameters/VehicleId'
      responses:
        '200':
          description: Vehículo encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Vehicle'

  # ─── TRIPS ───
  /trips:
    post:
      tags: [Trips]
      summary: Publicar viaje
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TripCreate'
      responses:
        '201':
          description: Viaje publicado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Trip'
    get:
      tags: [Trips]
      summary: Buscar viajes disponibles
      security: []
      parameters:
        - name: origin
          in: query
          required: true
          schema:
            type: string
          example: "Guatire"
        - name: destination
          in: query
          required: true
          schema:
            type: string
          example: "Caracas"
        - name: date
          in: query
          required: true
          schema:
            type: string
            format: date
          example: "2026-09-01"
        - name: passengers
          in: query
          schema:
            type: integer
            default: 1
        - name: lat
          in: query
          schema:
            type: number
            format: float
        - name: lng
          in: query
          schema:
            type: number
            format: float
        - name: radius_km
          in: query
          schema:
            type: integer
            default: 10
      responses:
        '200':
          description: Viajes encontrados
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TripSearchResult'
  /trips/{id}:
    get:
      tags: [Trips]
      summary: Detalle de viaje
      security: []
      parameters:
        - $ref: '#/components/parameters/TripId'
      responses:
        '200':
          description: Detalle del viaje
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Trip'
    patch:
      tags: [Trips]
      summary: Actualizar viaje (conductor)
      parameters:
        - $ref: '#/components/parameters/TripId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TripUpdate'
      responses:
        '200':
          description: Viaje actualizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Trip'
    delete:
      tags: [Trips]
      summary: Cancelar viaje (conductor)
      parameters:
        - $ref: '#/components/parameters/TripId'
      responses:
        '204':
          description: Viaje cancelado
  /trips/{id}/status:
    patch:
      tags: [Trips]
      summary: Cambiar estado del viaje (active, completed, cancelled)
      parameters:
        - $ref: '#/components/parameters/TripId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [status]
              properties:
                status:
                  type: string
                  enum: [active, completed, cancelled]
      responses:
        '200':
          description: Estado actualizado
  /trips/{id}/location:
    post:
      tags: [Trips]
      summary: Actualizar ubicación GPS del viaje en curso
      parameters:
        - $ref: '#/components/parameters/TripId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [lat, lng]
              properties:
                lat:
                  type: number
                  format: float
                lng:
                  type: number
                  format: float
                speed:
                  type: number
                  format: float
      responses:
        '204':
          description: Ubicación actualizada

  # ─── BOOKINGS ───
...
  /reviews/user/{userId}:
    get:
      tags: [Reviews]
      summary: Calificaciones recibidas por un usuario
      security: []
      parameters:
        - $ref: '#/components/parameters/UserId'
      responses:
        '200':
          description: Lista de calificaciones
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Review'

  # ─── INCIDENTS ───
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    UserId:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    VehicleId:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    TripId:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  schemas:
...
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        phone:
          type: string
        email:
          type: string
        id_doc:
          type: string
        role:
          type: string
          enum: [passenger, driver, admin]
        status:
          type: string
          enum: [active, suspended, blocked]
        rating:
          type: number
          format: float
          description: Calificación promedio
        completed_trips:
          type: integer
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    UserUpdate:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
          format: email
        selfie_url:
          type: string
...
    Vehicle:
      type: object
      properties:
        id:
          type: string
          format: uuid
        owner_id:
          type: string
          format: uuid
        plate:
          type: string
          example: "ABC-123"
        model:
          type: string
          example: "Toyota Corolla"
        year:
          type: integer
          example: 2018
        color:
          type: string
          example: "Blanco"
        seats:
          type: integer
          example: 4
        photos:
          type: array
          items:
            type: string
            description: S3 URL
        created_at:
          type: string
          format: date-time
    VehicleCreate:
      type: object
      required: [plate, model, year, seats]
      properties:
        plate:
          type: string
        model:
          type: string
        year:
          type: integer
        color:
          type: string
        seats:
          type: integer
        photos:
          type: array
          items:
            type: string
    Trip:
      type: object
      properties:
        id:
          type: string
          format: uuid
        driver_id:
          type: string
          format: uuid
        vehicle_id:
          type: string
          format: uuid
        origin:
          type: string
          example: "Guatire"
        destination:
          type: string
          example: "Caracas"
        origin_coords:
          $ref: '#/components/schemas/GeoPoint'
        destination_coords:
          $ref: '#/components/schemas/GeoPoint'
        departure_date:
          type: string
          format: date
        departure_time:
          type: string
          example: "06:30"
        seats_total:
          type: integer
        seats_available:
          type: integer
        price_per_seat_usd:
          type: number
          format: float
          example: 8.00
        price_per_seat_bs:
          type: number
          format: float
          nullable: true
        status:
          type: string
          enum: [scheduled, active, completed, cancelled]
        created_at:
          type: string
          format: date-time
    TripCreate:
      type: object
      required: [origin, destination, departure_date, departure_time, seats_total, price_per_seat_usd]
      properties:
        origin:
          type: string
        destination:
          type: string
        origin_lat:
          type: number
        origin_lng:
          type: number
        destination_lat:
          type: number
        destination_lng:
          type: number
        departure_date:
          type: string
          format: date
        departure_time:
          type: string
        seats_total:
          type: integer
        price_per_seat_usd:
          type: number
          format: float
        price_per_seat_bs:
          type: number
          format: float
          nullable: true
        vehicle_id:
          type: string
          format: uuid
    TripUpdate:
      type: object
      properties:
        departure_time:
          type: string
        seats_total:
          type: integer
        price_per_seat_usd:
          type: number
          format: float
    TripSearchResult:
      allOf:
        - $ref: '#/components/schemas/Trip'
        - type: object
          properties:
            driver_name:
              type: string
            driver_rating:
              type: number
              format: float
            driver_photo:
              type: string
            vehicle_plate:
              type: string
            vehicle_model:
              type: string
            distance_km:
              type: number
              format: float
              description: Distancia al origen del pasajero
...
    Review:
      type: object
      properties:
        id:
          type: string
          format: uuid
        trip_id:
          type: string
          format: uuid
        from_user_id:
          type: string
          format: uuid
        to_user_id:
          type: string
          format: uuid
        rating:
          type: integer
          minimum: 1
          maximum: 5
        comment:
          type: string
        created_at:
          type: string
          format: date-time
...
    GeoPoint:
      type: object
      properties:
        lat:
          type: number
          format: float
        lng:
          type: number
          format: float